changes to tests for multiple files
