+++ config_rooms/json/all_in_one.json
