+++ config_rooms/json/building_ref_roomfiles.json
@@ -1,0 +1,111 @@
+{
+"rooms": {
+        "room 1": "config_rooms/json/room_1.json",
+        "room 2": "config_rooms/json/room_2.json",
+        "room 3": "config_rooms/json/room_3.json",
+        "room 4": "config_rooms/json/room_4.json",
+        "room 5": "config_rooms/json/room_5.json",
+        "room 6": "config_rooms/json/room_6.json",
+        "room 7": "config_rooms/json/room_7.json",
+        "room 8": "config_rooms/json/room_8.json",
+        "room 9": "config_rooms/json/room_9.json"
+    },
+"global_settings":
+    {
+        "filename": "chem_mech/mcm_subset.fac",
+        "INCHEM_additional": true,
+        "particles": false,
+        "constrained_file": "constraints.csv",
+        "output_folder": "output/",
+        "dt": 0.002,
+        "H2O2_dep": true,
+        "O3_dep": false,
+        "custom": true,
+        "custom_filename": "custom_rxns.fac",
+        "diurnal": true,
+        "city": "London_urban",
+        "date": "21-06-2020",
+        "lat": 45.4,
+        "path": "/simulation/data",
+        "reactions_output": false,
+        "building_direction_in_radians": 1.57,
+        "air_density": 1.225,
+        "upwind_pressure_coefficient": 0.3,
+        "downwind_pressure_coefficient": -0.2
+  },
+"wind":{
+  "wind_speed": [
+    {"time": 0, "value": 2.32},
+    {"time": 3600, "value": 2.32},
+    {"time": 7200, "value": 1.42},
+    {"time": 10800, "value": 5.8},
+    {"time": 14400, "value": 5.4},
+    {"time": 18000, "value": 5.6},
+    {"time": 21600, "value": 3.9},
+    {"time": 25200, "value": 3.5},
+    {"time": 28800, "value": 3.3},
+    {"time": 32400, "value": 12.6},
+    {"time": 36000, "value": 14.5},
+    {"time": 39600, "value": 13.8},
+    {"time": 43200, "value": 16.0},
+    {"time": 46800, "value": 18.1},
+    {"time": 50400, "value": 12.4},
+    {"time": 54000, "value": 9.8},
+    {"time": 57600, "value": 9.0},
+    {"time": 61200, "value": 7.4},
+    {"time": 64800, "value": 7.6},
+    {"time": 68400, "value": 7.1},
+    {"time": 72000, "value": 11.8},
+    {"time": 75600, "value": 11.4},
+    {"time": 79200, "value": 11.5},
+    {"time": 82800, "value": 11.7}
+  ],
+  "wind_direction": [
+    {"time": 0, "value": 92.77},
+    {"time": 3600, "value": 92.77},
+    {"time": 7200, "value": 84.51},
+    {"time": 10800, "value": 27},
+    {"time": 14400, "value": 24},
+    {"time": 18000, "value": 24.5},
+    {"time": 21600, "value": 17},
+    {"time": 25200, "value": 15},
+    {"time": 28800, "value": 16},
+    {"time": 32400, "value": 275},
+    {"time": 36000, "value": 277},
+    {"time": 39600, "value": 281},
+    {"time": 43200, "value": 285.5},
+    {"time": 46800, "value": 284},
+    {"time": 50400, "value": 279},
+    {"time": 54000, "value": 131},
+    {"time": 57600, "value": 133},
+    {"time": 61200, "value": 132.5},
+    {"time": 64800, "value": 132},
+    {"time": 68400, "value": 130},
+    {"time": 72000, "value": 219},
+    {"time": 75600, "value": 221},
+    {"time": 79200, "value": 222},
+    {"time": 82800, "value": 220.5}
+  ],
+  "in_radians": false
+  },
+"apertures": [
+  {"room1": "room 1", "room2": "Front", "area": 0.000344},
+  {"room1": "room 1", "room2": "room 5", "area": 0.020998},
+  {"room1": "room 2", "room2": "Back", "area": 0.000874},
+  {"room1": "room 2", "room2": "room 5", "area": 0.011382},
+  {"room1": "room 3", "room2": "Back", "area": 0.000874},
+  {"room1": "room 3", "room2": "room 5", "area": 0.012031},
+  {"room1": "room 4", "room2": "Front", "area": 0.000983},
+  {"room1": "room 4", "room2": "Right", "area": 0.000983},
+  {"room1": "room 4", "room2": "room 5", "area": 0.010618},
+  {"room1": "room 6", "room2": "Front", "area": 0.000606},
+  {"room1": "room 6", "room2": "Left", "area": 0.000232},
+  {"room1": "room 6", "room2": "room 8", "area": 0.011924},
+  {"room1": "room 7", "room2": "Left", "area": 0.000232},
+  {"room1": "room 7", "room2": "Back", "area": 0.000232},
+  {"room1": "room 7", "room2": "room 8", "area": 0.013523},
+  {"room1": "room 8", "room2": "room 9", "area": 0.010618},
+  {"room1": "room 9", "room2": "Front", "area": 0.001129},
+  {"room1": "room 9", "room2": "Right", "area": 0.000232}
+  ]
+}
