+++ config_rooms/json/room_1.json
@@ -1,0 +1,96 @@
+{
+      "volume_in_m3": 10,
+      "surf_area_in_m2": 29.7,
+      "light_type": "Incand",
+      "glass_type": "glass_C",
+      "composition": {
+        "soft": 11.72,
+        "paint": 37.96,
+        "wood": 20.5,
+        "metal": 8.39,
+        "concrete": 0.44,
+        "paper": 2.12,
+        "lino": 0,
+        "plastic": 15.94,
+        "human": 0,
+        "glass": 2.93,
+        "other": 0
+      },
+      "temp_in_kelvin": {
+        "values": [
+          [0,290.62],
+          [3600,290.26],
+          [7200,289.91],
+          [10800,289.56],
+          [14400,289.2],
+          [18000,288.86],
+          [21600,288.5],
+          [25200,288.15],
+          [28800,289.01],
+          [32400,289.88],
+          [36000,290.73],
+          [39600,291.59],
+          [43200,292.46],
+          [46800,293.29],
+          [50400,294.15],
+          [54000,293.79],
+          [57600,293.44],
+          [61200,293.08],
+          [64800,292.73],
+          [68400,292.39],
+          [72000,292.03],
+          [75600,291.68],
+          [79200,291.32],
+          [82800,290.96]
+        ]
+      },
+      "rh_in_percent": {
+        "values": [
+          [0,50],
+          [82800,50]
+        ]
+      },
+      "airchange_in_per_second": {
+        "values": [
+          [0,1.39E-04],
+          [82800, 1.39E-04]
+        ]
+      },
+      "light_switch": {
+        "values": [
+          [0,0],
+          [21600,0],
+          [25200,1],
+          [64800,1],
+          [68400,0],
+          [82800,0]
+        ],
+        "continuous": false
+      },
+      "emissions": {
+        "LIMONENE": {
+          "values": [
+            { "start": 46800, "end": 50400, "value": 5.00E+08 }
+          ]
+        },
+        "BPINENE": {
+          "values": [
+            { "start": 46800, "end": 50400, "value": 5.00E+10}
+          ]
+        }
+      },
+      "n_adults": {
+        "values": [
+          { "time": 0, "value": 0 },
+          { "time": 82800, "value": 0 }
+        ],
+        "continuous": false
+      },
+      "n_children": {
+        "values": [
+          { "time": 0, "value": 0 },
+          { "time": 82800, "value": 0 }
+        ],
+        "continuous": false
+      }
+    }
