+++ config_rooms/json/room_3.json
@@ -1,0 +1,192 @@
+{
+      "volume_in_m3": 32.4,
+      "surf_area_in_m2": 61.8,
+      "light_type": "UFT",
+      "glass_type": "low_emissivity",
+      "composition": {
+        "soft": 20,
+        "paint": 20,
+        "wood": 20,
+        "metal": 0,
+        "concrete": 10,
+        "paper": 30,
+        "lino": 0,
+        "plastic": 0,
+        "human": 0,
+        "glass": 0,
+        "other": 0
+      },
+      "temp_in_kelvin": {
+        "values": [
+          [0,290],
+          [3600,290],
+          [7200,290],
+          [10800,290],
+          [14400,290],
+          [18000,296],
+          [21600,296],
+          [25200,296],
+          [28800,291],
+          [32400,291],
+          [36000,291],
+          [39600,291],
+          [43200,291],
+          [46800,291],
+          [50400,291],
+          [54000,291],
+          [57600,294],
+          [61200,294],
+          [64800,294],
+          [68400,294],
+          [72000,294],
+          [75600,294],
+          [79200,294],
+          [82800,294]
+        ]
+      },
+      "rh_in_percent": {
+        "values": [
+          [0,70],
+          [3600,70],
+          [7200,70],
+          [10800,70],
+          [14400,70],
+          [18000,70],
+          [21600,70],
+          [25200,65],
+          [28800,65],
+          [32400,65],
+          [36000,65],
+          [39600,65],
+          [43200,65],
+          [46800,65],
+          [50400,65],
+          [54000,65],
+          [57600,80],
+          [61200,80],
+          [64800,80],
+          [68400,80],
+          [72000,80],
+          [75600,80],
+          [79200,80],
+          [82800,80]
+        ]
+      },
+      "airchange_in_per_second": {
+        "values": [
+          [0,6.94E-05],
+          [3600,6.94E-05],
+          [7200,6.94E-05],
+          [10800,6.94E-05],
+          [14400,6.94E-05],
+          [18000,6.94E-05],
+          [21600,6.94E-05],
+          [25200,2.78E-04],
+          [28800,2.78E-04],
+          [32400,2.78E-04],
+          [36000,2.78E-04],
+          [39600,2.78E-04],
+          [43200,2.78E-04],
+          [46800,2.78E-04],
+          [50400,2.78E-04],
+          [54000,2.78E-04],
+          [57600,5.56E-05],
+          [61200,5.56E-05],
+          [64800,5.56E-05],
+          [68400,5.56E-05],
+          [72000,5.56E-05],
+          [75600,5.56E-05],
+          [79200,5.56E-05],
+          [82800,5.56E-05]
+        ]
+      },
+      "light_switch": {
+        "values": [
+          [0,0],
+          [3600,0],
+          [7200,0],
+          [10800,0],
+          [14400,0],
+          [18000,1],
+          [21600,1],
+          [25200,1],
+          [28800,0],
+          [32400,0],
+          [36000,0],
+          [39600,0],
+          [43200,0],
+          [46800,0],
+          [50400,0],
+          [54000,0],
+          [57600,1],
+          [61200,1],
+          [64800,1],
+          [68400,1],
+          [72000,1],
+          [75600,1],
+          [79200,1],
+          [82800,1]
+        ],
+        "continuous": false
+      },
+      "emissions": {},
+      "n_adults": {
+        "values": [
+          [0,2],
+          [3600,2],
+          [7200,2],
+          [10800,2],
+          [14400,2],
+          [18000,2],
+          [21600,0],
+          [25200,0],
+          [28800,0],
+          [32400,0],
+          [36000,0],
+          [39600,0],
+          [43200,0],
+          [46800,0],
+          [50400,0],
+          [54000,0],
+          [57600,2],
+          [61200,2],
+          [64800,2],
+          [68400,2],
+          [72000,2],
+          [75600,2],
+          [79200,2],
+          [82800,2]
+        ],
+        "continuous": false
+      },
+      "n_children": {
+        "values": [
+          [0,1],
+          [3600,1],
+          [7200,1],
+          [10800,1],
+          [14400,1],
+          [18000,1],
+          [21600,1],
+          [25200,1],
+          [28800,0],
+          [32400,0],
+          [36000,0],
+          [39600,0],
+          [43200,0],
+          [46800,0],
+          [50400,1],
+          [54000,1],
+          [57600,1],
+          [61200,1],
+          [64800,1],
+          [68400,1],
+          [72000,1],
+          [75600,1],
+          [79200,1],
+          [82800,1]
+        ],
+        "continuous": false
+      }
+    
+}
