+++ config_rooms/json/room_4.json
@@ -1,0 +1,59 @@
+{
+      "volume_in_m3": 40.5,
+      "surf_area_in_m2": 81.3,
+      "light_type": "CFT",
+      "glass_type": "low_emissivity_film",
+      "composition": {
+        "soft": 20,
+        "paint": 40,
+        "wood": 30,
+        "metal": 5,
+        "concrete": 0,
+        "paper": 0,
+        "lino": 0,
+        "plastic": 0,
+        "human": 0,
+        "glass": 5,
+        "other": 0
+      },
+      "temp_in_kelvin": {
+        "values": [
+          [0,293],
+          [82800, 293]
+        ]   
+      },
+      "rh_in_percent": {
+        "values": [
+          [0,50],
+          [82800,50]
+        ]
+      },
+      "airchange_in_per_second": {
+        "values": [
+          [0,0],
+          [82800, 0]
+        ]
+      },
+      "light_switch": {
+        "values": [
+          [0,1],
+          [82800, 1]
+        ],
+        "continuous": false
+      },
+      "emissions": {},
+      "n_adults": {
+        "values": [
+          { "time": 0, "value": 2 },
+          { "time": 82800, "value": 2 }
+        ],
+        "continuous": false
+      },
+      "n_children": {
+        "values": [
+          { "time": 0, "value": 1 },
+          { "time": 82800, "value": 1 }
+        ],
+        "continuous": false
+      }
+    }
