+++ config_rooms/json/room_7.json
@@ -1,0 +1,60 @@
+{
+      "volume_in_m3": 22.5,
+      "surf_area_in_m2": 48,
+      "light_type": "Incand",
+      "glass_type": "low_emissivity_film",
+      "composition": {
+        "soft": 0,
+        "paint": 45,
+        "wood": 40,
+        "metal": 0,
+        "concrete": 10,
+        "paper": 0,
+        "lino": 0,
+        "plastic": 5,
+        "human": 0,
+        "glass": 0,
+        "other": 0
+      },
+      "temp_in_kelvin": {
+        "values": [
+          [0,293],
+          [82800, 293]
+        ]    
+      },
+      "rh_in_percent": {
+        "values": [
+          [0,50],
+          [82800,50]
+        ]
+      },
+      "airchange_in_per_second": {
+        "values": [
+          [0,0],
+          [82800, 0]
+        ]
+      },
+      "light_switch": {
+        "values": [
+          [0,1],
+          [82800, 1]
+        ],
+        "continuous": false
+      },
+      "emissions": {},
+      "n_adults": {
+        "values": [
+          { "time": 0, "value": 2 },
+          { "time": 82800, "value": 2 }
+        ],
+        "continuous": false
+      },
+      "n_children": {
+        "values": [
+          { "time": 0, "value": 1 },
+          { "time": 82800, "value": 1 }
+        ],
+        "continuous": false
+      }
+    
+}
