+++ config_rooms/json/room_8.json
@@ -1,0 +1,60 @@
+{
+      "volume_in_m3": 50.4,
+      "surf_area_in_m2": 86.4,
+      "light_type": "UFT",
+      "glass_type": "glass_C",
+      "composition": {
+        "soft": 0,
+        "paint": 22,
+        "wood": 0,
+        "metal": 0,
+        "concrete": 10,
+        "paper": 0,
+        "lino": 50,
+        "plastic": 18,
+        "human": 0,
+        "glass": 0,
+        "other": 0
+      },
+      "temp_in_kelvin": {
+        "values": [
+          [0,293],
+          [82800, 293]
+        ]    
+      },
+      "rh_in_percent": {
+        "values": [
+          [0,50],
+          [82800,50]
+        ]
+      },
+      "airchange_in_per_second": {
+        "values": [
+          [0,0],
+          [82800, 0]
+        ]
+      },
+      "light_switch": {
+        "values": [
+          [0,1],
+          [82800, 1]
+        ],
+        "continuous": false
+      },
+      "emissions": {},
+      "n_adults": {
+        "values": [
+          { "time": 0, "value": 0 },
+          { "time": 82800, "value": 0 }
+        ],
+        "continuous": false
+      },
+      "n_children": {
+        "values": [
+          { "time": 0, "value": 1 },
+          { "time": 82800, "value": 1 }
+        ],
+        "continuous": false
+      }
+    
+}
