+++ config_rooms/json/room_9.json
@@ -1,0 +1,59 @@
+{
+      "volume_in_m3": 132,
+      "surf_area_in_m2": 158,
+      "light_type": "Incand",
+      "glass_type": "no_sunlight",
+      "composition": {
+        "soft": 18,
+        "paint": 35,
+        "wood": 33,
+        "metal": 11,
+        "concrete": 0,
+        "paper": 0,
+        "lino": 0,
+        "plastic": 0,
+        "human": 0,
+        "glass": 3,
+        "other": 0
+      },
+      "temp_in_kelvin": {
+        "values": [
+          [0,293],
+          [82800, 293]
+        ]    
+      },
+      "rh_in_percent": {
+        "values": [
+          [0,50],
+          [82800,50]
+        ]
+      },
+      "airchange_in_per_second": {
+        "values": [
+          [0,0],
+          [82800, 0]
+        ]
+      },
+      "light_switch": {
+        "values": [
+          [0,1],
+          [82800, 1]
+        ],
+        "continuous": false
+      },
+      "emissions": {},
+      "n_adults": {
+        "values": [
+          { "time": 0, "value": 0 },
+          { "time": 82800, "value": 0 }
+        ],
+        "continuous": false
+      },
+      "n_children": {
+        "values": [
+          { "time": 0, "value": 1 },
+          { "time": 82800, "value": 1 }
+        ],
+        "continuous": false
+      }
+    }
